// types.ts
export interface SKU {
  id: string;
  name: string;
  rawIngredients: number;
  primaryPackaging: number;
  secondaryPackaging: number;
  tollProcessing: number;
  inboundFreight: number;
  unitsPerCase: number;
  casesPerPallet: number;
  volumeMixPct: number; // what % of total company volume is this SKU (0-1, all SKUs should sum to 1)
}

export interface SKULibraryState {
  skus: SKU[];
}